requirements.txt
Flask==2.1.2
Pillow==9.0.1
numpy==1.21.2
opencv-python==4.5.3.20210927
scikit-image==0.18.3